WildlandStorageManifest(): don't take 'parent' on creation
A storage manifest doesn't need to know its parent -- only the parent needs to know which storage it is assigned to. This change makes the whole process of actor/container/storage manifest much more natural and easier to understand.

Also renamed WildlandManifest's add_storage_manifest() to add_storage_backend(), as it should be. This introduced incompatibility in our API, so requires updates to example scripts. Trivial though.